ingest/uploader/requirements.txt
audio-metadata==0.11.1
minio==7.1.7
pyinstaller==5.0.1
pyparsing==2.4.7
PyQt5==5.15.6
psycopg2-binary==2.8.6
tqdm==4.64.0
